+++ src/main.rs
@@ -8,7 +8,7 @@ use std::{
 };
 
 // Constants
-const VERSION: &str = "0.0.7";
+const VERSION: &str = "0.0.8";
 
 // Utility functions
 
@@ -26,12 +26,11 @@ fn check_tool(tool: &str) {
 // Function for creating folders and parents
 fn make_dirs(dir: &str) -> std::io::Result<()> {
     let result = create_dir_all(dir);
-    match result {
-        Err(e) => {
-            eprintln!("error: {}", e);
-            exit(2);
-        }
-        Ok(_) => Ok(()),
+    if let Err(e) = result {
+        eprintln!("error: {}", e);
+        exit(2);
+    } else {
+        Ok(())
     }
 }
 
@@ -39,12 +38,11 @@ fn make_dirs(dir: &str) -> std::io::Result<()> {
 fn make_file(file: &str, content: String) -> std::io::Result<()> {
     let mut file = File::create(file)?;
     let result = file.write_all(&content.as_bytes());
-    match result {
-        Err(e) => {
-            eprintln!("error: {}", e);
-            exit(3);
-        }
-        Ok(_) => Ok(()),
+    if let Err(e) = result {
+        eprintln!("error: {}", e);
+        exit(3);
+    } else {
+        Ok(())
     }
 }
 
@@ -95,12 +93,9 @@ fn prj_name() -> String {
     let project = format!("{root}/{package}");
     // Make directories structure
     let dir_ret = make_dirs(format!("{project}").as_str());
-    match dir_ret {
-        Err(e) => {
-            eprintln!("error: {}", e);
-            exit(4);
-        }
-        Ok(_) => (),
+    if let Err(e) = dir_ret {
+        eprintln!("error: {}", e);
+        exit(4);
     }
     // Make file structures
     let file_ret = make_file(
@@ -114,12 +109,9 @@ fn prj_name() -> String {
 "
         .to_string(),
     );
-    match file_ret {
-        Err(e) => {
-            eprintln!("error: {}", e);
-            exit(4);
-        }
-        Ok(_) => (),
+    if let Err(e) = file_ret {
+        eprintln!("error: {}", e);
+        exit(4);
     }
     name
 }
@@ -177,12 +169,11 @@ docs/_build/
 /site"
                 .to_string(),
         );
-        let ret = match file_ret {
-            Err(e) => {
-                eprintln!("error: {}", e);
-                exit(5);
-            }
-            Ok(_) => true,
+        let ret = if let Err(e) = file_ret {
+            eprintln!("error: {}", e);
+            exit(5);
+        } else {
+            true
         };
         return ret;
     }
@@ -195,12 +186,9 @@ fn prj_test(name: &str) {
     if confirm {
         // Make directories structure
         let dir_ret = make_dirs(format!("{name}/tests").as_str());
-        match dir_ret {
-            Err(e) => {
-                eprintln!("error: {}", e);
-                exit(6);
-            }
-            Ok(_) => (),
+        if let Err(e) = dir_ret {
+            eprintln!("error: {}", e);
+            exit(6);
         }
         // Make file structures
         let init_file = make_file(
@@ -213,12 +201,9 @@ fn prj_test(name: &str) {
 "
             .to_string(),
         );
-        match init_file {
-            Err(e) => {
-                eprintln!("error: {}", e);
-                exit(6);
-            }
-            Ok(_) => (),
+        if let Err(e) = init_file {
+            eprintln!("error: {}", e);
+            exit(6);
         }
         let project_name = name.to_lowercase();
         let all_module = make_file(
@@ -248,12 +233,9 @@ if __name__ == '__main__':
             )
             .to_string(),
         );
-        match all_module {
-            Err(e) => {
-                eprintln!("error: {}", e);
-                exit(6);
-            }
-            Ok(_) => (),
+        if let Err(e) = all_module {
+            eprintln!("error: {}", e);
+            exit(6);
         }
     }
 }
@@ -347,12 +329,9 @@ changelog = ''
     );
     // Write pyproject.toml
     let pyproject = make_file(format!("{name}/pyproject.toml").as_str(), content);
-    match pyproject {
-        Err(e) => {
-            eprintln!("error: {}", e);
-            exit(7);
-        }
-        Ok(_) => (),
+    if let Err(e) = pyproject {
+        eprintln!("error: {}", e);
+        exit(7);
     }
 }
 
@@ -388,19 +367,13 @@ script:
   - pytest tests"
             ),
         );
-        match travis {
-            Err(e) => {
-                eprintln!("error: {}", e);
-            }
-            Ok(_) => (),
+        if let Err(e) = travis {
+            eprintln!("error: {}", e);
         }
     } else if ci.as_str() == "CircleCI" {
         let dir_ret = make_dirs(format!("{name}/.circleci").as_str());
-        match dir_ret {
-            Err(e) => {
-                eprintln!("error: {}", e);
-            }
-            Ok(_) => (),
+        if let Err(e) = dir_ret {
+            eprintln!("error: {}", e);
         }
         let circle = make_file(
             format!("{name}/.circleci/config.yml").as_str(),
@@ -431,11 +404,8 @@ jobs:
 "
             ),
         );
-        match circle {
-            Err(e) => {
-                eprintln!("error: {}", e);
-            }
-            Ok(_) => (),
+        if let Err(e) = circle {
+            eprintln!("error: {}", e);
         }
     }
 }
@@ -480,18 +450,12 @@ fn prj_remote(name: &str) {
                 remote.to_lowercase()
             );
             let dir_ret = make_dirs(issue_folder.as_str());
-            match dir_ret {
-                Err(e) => {
-                    eprintln!("error: {}", e);
-                }
-                Ok(_) => (),
+            if let Err(e) = dir_ret {
+                eprintln!("error: {}", e);
             }
             let dir_ret = make_dirs(merge_folder.as_str());
-            match dir_ret {
-                Err(e) => {
-                    eprintln!("error: {}", e);
-                }
-                Ok(_) => (),
+            if let Err(e) = dir_ret {
+                eprintln!("error: {}", e);
             }
             let feature_content = format!(
                 "## Description
@@ -512,11 +476,8 @@ Additional context
                 format!("{issue_folder}/feature.md").as_str(),
                 feature_content,
             );
-            match feature_issue {
-                Err(e) => {
-                    eprintln!("error: {}", e);
-                }
-                Ok(_) => (),
+            if let Err(e) = feature_issue {
+                eprintln!("error: {}", e);
             }
             let bug_content = format!(
                 "## Description of problem
@@ -542,11 +503,8 @@ Additional context
                 name.to_lowercase()
             );
             let bug_issue = make_file(format!("{issue_folder}/bug.md").as_str(), bug_content);
-            match bug_issue {
-                Err(e) => {
-                    eprintln!("error: {}", e);
-                }
-                Ok(_) => (),
+            if let Err(e) = bug_issue {
+                eprintln!("error: {}", e);
             }
             let merge_content = format!(
                 "## What does this MR do and why?
@@ -566,40 +524,28 @@ Numbered steps to set up and validate the change are strongly suggested.
                 name.to_lowercase()
             );
             let merge_issue = make_file(format!("{merge_folder}/merge.md").as_str(), merge_content);
-            match merge_issue {
-                Err(e) => {
-                    eprintln!("error: {}", e);
-                }
-                Ok(_) => (),
+            if let Err(e) = merge_issue {
+                eprintln!("error: {}", e);
             }
         // Github
         } else if remote.as_str() == "Github" {
             let issue_folder = format!("{}/.{}/ISSUE_TEMPLATE", name, remote.to_lowercase());
             let merge_folder = format!("{}/.{}/PULL_REQUEST_TEMPLATE", name, remote.to_lowercase());
             let dir_ret = make_dirs(issue_folder.as_str());
-            match dir_ret {
-                Err(e) => {
-                    eprintln!("error: {}", e);
-                }
-                Ok(_) => (),
+            if let Err(e) = dir_ret {
+                eprintln!("error: {}", e);
             }
             let dir_ret = make_dirs(merge_folder.as_str());
-            match dir_ret {
-                Err(e) => {
-                    eprintln!("error: {}", e);
-                }
-                Ok(_) => (),
+            if let Err(e) = dir_ret {
+                eprintln!("error: {}", e);
             }
             let config_content = "blank_issues_enabled: false\n".to_string();
             let config_file = make_file(
                 format!("{issue_folder}/config.yml").as_str(),
                 config_content,
             );
-            match config_file {
-                Err(e) => {
-                    eprintln!("error: {}", e);
-                }
-                Ok(_) => (),
+            if let Err(e) = config_file {
+                eprintln!("error: {}", e);
             }
             let feature_content = format!(
                 "name: Feature Request
@@ -642,11 +588,8 @@ body:
                 format!("{issue_folder}/feature.yml").as_str(),
                 feature_content,
             );
-            match feature_issue {
-                Err(e) => {
-                    eprintln!("error: {}", e);
-                }
-                Ok(_) => (),
+            if let Err(e) = feature_issue {
+                eprintln!("error: {}", e);
             }
             let bug_content = format!(
                 "name: Bug Report
@@ -700,11 +643,8 @@ body:
                 name.to_lowercase()
             );
             let bug_issue = make_file(format!("{issue_folder}/bug.yml").as_str(), bug_content);
-            match bug_issue {
-                Err(e) => {
-                    eprintln!("error: {}", e);
-                }
-                Ok(_) => (),
+            if let Err(e) = bug_issue {
+                eprintln!("error: {}", e);
             }
             let merge_content = format!(
                 "# Title of Pull Request
@@ -731,12 +671,144 @@ assignees: {}
                 format!("{merge_folder}/pull_request_template.md").as_str(),
                 merge_content,
             );
-            match merge_issue {
-                Err(e) => {
-                    eprintln!("error: {}", e);
-                }
-                Ok(_) => (),
+            if let Err(e) = merge_issue {
+                eprintln!("error: {}", e);
             }
+        }
+    }
+}
+
+// Project tox
+fn prj_tox(name: &str, venv: bool) {
+    // Create tox ini
+    let confirm = prompt_confirm("Do you want to configure tox?", false, "None");
+    if confirm {
+        let mut pip = std::process::Command::new("pip3");
+        // Activate venv
+        if venv {
+            pip.env("PATH", "venv/bin");
+        }
+        let output = pip
+            .arg("install")
+            .arg("--timeout=10")
+            .arg("--retries=1")
+            .arg("tox")
+            .current_dir(&name)
+            .output()
+            .expect("pip should be installed");
+        // Check if command exit successfully
+        if !output.status.success() {
+            eprintln!("error: tox installation failed");
+            return;
+        }
+        // Write tox.ini
+        let tox_ini_content = "[tox]
+envlist = py310, py311, py312
+isolated_build = True
+
+[testenv]
+labels = test, core
+deps = pytest
+commands = pytest tests"
+            .to_string();
+        let tox_ini = make_file(format!("{name}/tox.ini").as_str(), tox_ini_content);
+        if let Err(e) = tox_ini {
+            eprintln!("error: {}", e);
+        }
+    }
+}
+
+fn prj_docs(name: &str, venv: bool) {
+    let options = vec!["None", "Sphinx", "MKDocs"];
+    let docs = prompt_select("Select document generator:", options, "None");
+    let docs_home = format!("{name}/docs");
+    // Create docs folder
+    let docs_folder = make_dirs(format!("{docs_home}").as_str());
+    if let Err(e) = docs_folder {
+        eprintln!("error: {}", e);
+        exit(4);
+    }
+    if docs.as_str() == "Sphinx" {
+        // Install sphinx
+        let mut pip = std::process::Command::new("pip3");
+        // Activate venv
+        if venv {
+            pip.env("PATH", "venv/bin");
+        }
+        let output = pip
+            .arg("install")
+            .arg("--timeout=10")
+            .arg("--retries=1")
+            .arg("sphinx")
+            .current_dir(&name)
+            .output()
+            .expect("pip should be installed");
+        // Check if command exit successfully
+        if !output.status.success() {
+            eprintln!("error: sphinx installation failed");
+            return;
+        }
+        // Start documentation;
+        let sphinx_bin = format!("../venv/bin/sphinx-quickstart");
+        let mut sphinx_quickstart = std::process::Command::new(sphinx_bin);
+        // Activate venv
+        if venv {
+            sphinx_quickstart.env("PATH", "venv/bin");
+        }
+        let output = sphinx_quickstart
+            .arg("--quiet")
+            .arg("--sep")
+            .arg(format!("--project={}", name.to_lowercase()))
+            .arg("--author=''")
+            .arg("-v='0.0.1'")
+            .arg("--ext-autodoc")
+            .arg("--ext-doctest")
+            .arg("--ext-viewcode")
+            .arg("--makefile")
+            .current_dir(docs_home)
+            .output()
+            .expect("sphinx-quickstart should be installed");
+        // Check if command exit successfully
+        if !output.status.success() {
+            eprintln!("error: sphinx documentation creation failed");
+        }
+    } else if docs.as_str() == "MKDocs" {
+        // Install mkdocs
+        let mut pip = std::process::Command::new("pip3");
+        // Activate venv
+        if venv {
+            pip.env("PATH", "venv/bin");
+        }
+        let output = pip
+            .arg("install")
+            .arg("--timeout=10")
+            .arg("--retries=1")
+            .arg("mkdocs")
+            .current_dir(&name)
+            .output()
+            .expect("pip should be installed");
+        // Check if command exit successfully
+        if !output.status.success() {
+            eprintln!("error: mkdocs installation failed");
+            return;
+        }
+        // Start documentation;
+        let mkdocs_bin = format!("venv/bin/mkdocs");
+        let mut mkdocs_new = std::process::Command::new(mkdocs_bin);
+        // Activate venv
+        if venv {
+            mkdocs_new.env("PATH", "venv/bin");
+        }
+        let output = mkdocs_new
+            .arg("new")
+            .arg("--quiet")
+            .arg(".")
+            .current_dir(&name)
+            .output()
+            .expect("mkdocs should be installed");
+        // Check if command exit successfully
+        if !output.status.success() {
+            eprintln!("error: mkdocs documentation creation failed");
         }
     }
 }
@@ -767,6 +839,10 @@ fn main() {
     }
     // Write pyproject.toml
     prj_toml(&name, &deps);
+    // Tox
+    prj_tox(&name, venv);
+    // Documentation
+    prj_docs(&name, venv);
     // Finish scaffolding process
     println!("Project `{name}` created")
 }